feat(selection): add SelectionContext dataclass
Coin selection requires pre computations that usually are common to all
algorithms. Also, the combination of parameters is unique for each coin
selection.

This wrapper was created, to abstract and provide the common context
that each algorithm expects and to attach the finally produced
transaction to the unique environment where it was constructed.