feat(metrics): add Waste metric
This implementations considers waste as the sum of two values:

waste = timing_cost + creation_cost

Timing cost is the cost associated to the current fee rate and some long
term fee rate used as a treshold to consolidate UTXOs.

timing_cost = txin_size * current_fee_rate - txin_size * long_term_fee_rate

Timing cost can be negative if the current_fee_rate is cheaper than the
long_term_fee_rate, or zero if they are equal.

The name of this cost is derived from the action of making bets against
some kind of market (in this case the fee market) to gain profits using
the interpretation of the future behaviour of that market based on a
partial information game.

Creation cost is the cost associated to the surplus of coins besides the
transaction amount and transaction fees. It can happen in the form of a
change output or in the form of excess fees paid to the miner.

Change cost is derived from the cost of adding the extra output to the
transaction and spending that output in the future.

change_cost = current_fee_rate * change_output_size + long_term_feerate * change_spend_size

Excess happens when there is not change, and the surplus of coins is
spend as part of the fees to the miner:

excess = tx_total_value - tx_fees - target

Where target is the desired amount to send.

Creation cost can be zero if there is a perfect match as result of the coin
selection algorithm.

So, waste can be zero or negative if the creation cost is zero and the timing
cost is less than or equal to zero

Later this can be used in the wallet to select the most optimal coin
selection algorithm.