refactor(selection): add base_fee as property
Although the most common target for a selection algorithm using fee
rated utxos is the payment amount plus the fees of the transaction
template including them; some other algorithms don't use this exact
amount, but need the fees to calculate a reasonable target.

This commit provides this amount adding the "base_fee" property to the
SelectionContext class.