feat(scenario): add scenarios generator command
A scenario is a CSV file of the format: <block-id>,<amount>,<fee-rate>.

Block id is a parameter used to force a group of outgoing payments to
participate in the same transaction.

The amount can be positive (incoming) or negative (outgoing).
Along with the fee rate, it is represented in terms of bitcoin, with 8
decimal places for the remaining satoshis.

The fee rate is expressed in terms of bitcoin per kilobyte.