fix(selection): sort fee rated utxos in context
The order between plain UTxOs and the collection that stores its
effective value may change as the effective value is mainly driven by
the fees paid by each UTxO, which is tightly tied to the weight of them.

Currently we are using UTxOs with the same weight, but that may change
in the future and cause hard to trace errors.