refactor(simulation): add Simulation class
The functions to run simulations grew with time and became more complex,
with many different branches possible in each execution.

The idea of implementing a Simulation class is to provide a context
manager to handle the simulations more simply, abstracting different
parts of the simulation control in their respective methods and at the
same time, allowing enough flexibility to expand the functionality in
the future.