fix(selection): add handlers for PulpSolverError
In some environments, Pulp can fail due to unexplained causes. In some
cases a number of retries can fix those failures.

This commit implements a callable class using the exponential backoff
retrial algorithm, also including a way to record in a JSON file the
model in execution at the moment of the failure.