feat(simulation): split simulation from dir setup
The former Simulation class handled the setup of the directories,
iterating over scenario-model pairs and launching each simulation. It
also performed other related actions, such as filtering some
scenarios/model combinations.

The simulation method involved different heterogeneous tasks and grew
each time a new fix or feat appeared. In the end, we were not talking
about a single simulation, but a suite of them run together thanks to
the Simulation class.

It seemed natural to split that class into SimulationBench and a new,
more specific, Simulation class. The former involving the setup of the
directory of the simulation, the filtering, and the dispatch of
scenario-model pairs, and the second one, focusing on the details of a
single simulation (one scenario and one model), separating the selection
task into many different steps.