feat(simulation): filter log keys with value zero
Log info entries during the execution of the simulation mainly have the
objective of allowing the user to track the progress and know if the
models are working as designed.

Some key-value pairs included in those log entries usually have a zero
value, not adding much information but taking screen space when logging
the transaction entry.

The remotion of those uninformative pairs will allow better use of the
screen space and retrieve a more informational output for the user.