fix(selection): rule out negative change in greatest_first
The `greatest_first` selection model had a bug in the computation of the
overpayment. While performing the selection of the minimal number of
utxos to pay for the payment request, it used the fee-rated utxo values.
However, at the moment of getting the overpayment, the used amount was
the whole value of the utxo, without the fee discount, producing greater
values than the real ones.
Finally, at the moment of fixing the rounding errors produced during the
selection, the logic of the function obtained sometimes negative values
due to this miscalculation in the overpayment.
Thankfully, the negative amount was reflected in the change in certain
selections, revealing the bug.