chore(data): add simulation started on 01/09/2023
This commit adds the data results generated by the simulation done with
the version of the repo at commit d0f0b35, which means it includes among
other changes, the following:

- 9879fa9: fix of the greatest-first change creation error.
- 39aad1b: fix of PulpSolverError (add retry decorator).
- d0f0b35: addition of `maximize-effective-value-with-fallback` selection
  model.

The data uploaded with this commit is missing the results of the
`avoid-change`, `minimize-waste`, `maximize-effective-value`, and
`maximize-effective-value-with-fallback` models for the `random_blocks`
scenario due to a sudden change in the execution environment. Those will
be added in a subsequent commit.