fix(selection): use CONSOLIDATION_FEE_RATE for change input cost
Change cost is an aproximate way to calculate how much the cost of
creating a change UTxO will be. The cost of that UTxO as output is easy
to establish as it is the weight of it times the fee rate estimated at
the moment of producing the transaction. The cost of the UTxO as input,
on the other hand, is harder to estimate, as it involves market dynamics
that we cannot know beforehand.

Because of that, we use the CONSOLIDATION_FEE_RATE as a fee rate
reference, to establish a minimum cost the use of a change output will
have as input.