fix(selection): add missing defaults for tx digest data
When the transaction fails and there is no final transaction to extract
data from, the tuple that will be written in the csv still needs the
missing values to avoid the misalignment of the data.

If the transaction fails, the final tuple needs 9 default values for the
transaction data. Some defaults were missing and the resulting data was
corrupted.

This commits adds the missing default values fixing the error.