fix(utils): replace ROUND_UP by ROUND_DOWN
This is a serious issue, mitigated by the transaction validity check
performed before the logging of transactions.

Because of the error that floating point arithmetic may introduce,
quantities processed in the simulations where rounded up to the amount
the unit represented needed (eight decimals in bitcoins, no decimals in
satoshis).

However, we were doing that rounding UP, making any quantities less than
a bitcoin or a satoshi, a whole new unit out of nothing.

This change discards any quantities below the threshold and avoids the
creation of new units by rounding DOWN.